Fix the floating point crash
If the double is greater than what the int can hold, floating point
error is genereted.
I doubt that anything that far away is too importnat.
that's why the return type remains an int, but the double is casted to
a long

+++ src/perlin.cpp
@@ -53,7 +53,7 @@
  * Eliot
  */
 
-inline int fastfloor(const double x) { return int(x > 0 ? x : x - 1); }
+inline int fastfloor(const double x) { return long(x > 0 ? x : x - 1); }
 
 //static double dot( const int* g, const double x, const double y ) { return g[0]*x + g[1]*y; }
 inline double dot(const double *g, const double x, const double y, const double z) { return g[0] * x + g[1] * y + g[2] * z; }
